Added and fixed methods in SimpleGameObject, added default graphic to GraphicsHelper, fixed alpha blending issue.
setGraphic methods in SimpleGameObject have been improved.
getRenderSystem was added. This is useful for things like changing the
color of the object's layer.

GraphicsHelper now has function *getDefaultGraphic* to return an
instance of a default graphic.

Previously, changing the alpha value for a layer did not work correctly.
It has been fixed.