ShadeRenderer, ParentAssignmentHandler, FlashingEffect, getCameraEdge changes, other small things
Add a ShadeRenderer to render a solid color over the whole screen on a
specific layer. ParentAssignmentHandler allows a component to know when
it was added to an entity and get a reference to that entity.
FlashingEffect can make QuadRenderSystem or ShadeRenderer to flash on a
specific layer. The getCamera_Edge() methods now return numbers in grid
units rather than pixels.