Add features, improve and simplify QuadRenderSystem
Added ScrollingGraphicArea and ScrollingImage for convenient scrolling
backgrounds. Jumpy Bug has been updated to demonstrate usage.

Simplified QuadRenderSystem by creating the Quad interface. Any Entity
can now be a Quad by implementing the new interface and can be added to
QuadRenderSystem.

Also changed the addGraphic method to getGraphic because it makes more
sense that way (sometimes an existing Graphic is returned instead of
making a new one). The addGraphic method is still there but it is marked
as deprecated and just calls the new method.

setParameters was also changed in GraphicsHelper. It now takes a boolean
'repeating'. That can be used to make Graphics repeat instead of
clamping to the edge.